requirements.txt
flask_mysqldb
mysql-connector-python